Feature/bc performance issue 77

+++ clean.sh
@@ -5,14 +5,24 @@ docker rm -f $(docker ps -aq)
 #TODO [ "${DOCKER_MACHINE_NAME}" == "orderer" ]  && EXECUTE_BY_ORDERER=1 runCLIWithComposerOverrides down || runCLIWithComposerOverrides down
 
 docker volume prune -f
+
 docker rmi -f $(docker images -q -f "reference=dev-*")
 
-if [ -z "$DOCKER_HOST" ] ; then
-     sudo rm -rf crypto-config/
-     sudo rm -rf data/
-else
-    docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
-    docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
+if [ -n "${FORCE_CLEAN_INSTALL}" ]; then
+
+    echo "Cleaning because FORCE_CLEAN_INSTALL is set to ${FORCE_CLEAN_INSTALL}"
+
+    docker volume rm -f peer0-volume
+    docker volume rm -f peer1-volume
+
+    if [ -z "$DOCKER_HOST" ] ; then
+         sudo rm -rf crypto-config/
+         sudo rm -rf data/
+    else
+        docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
+        docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
+    fi
+
 fi
 
 #docker rmi -f $(docker images -q -f "reference=olegabu/fabric-starter-client")
